Make sizoscope-cli a regular dotnet tool
Move the CLI diff logic directly into dotnet-tool-cli instead of
launching a separate AOT-compiled exe via Process.Start. The console
tool doesn't need the shim pattern that the WinForms GUI requires.

Remove the standalone sizoscope-cli project and its CI build steps.
