+++ dotnet-tool-cli/sizoscope-cli.csproj
@@ -3,6 +3,8 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <RollForward>Major</RollForward>
     <Version Condition="'$(Version)' == ''">42.42.42.42</Version>
 
@@ -17,13 +19,24 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="TurboXml" Version="2.0.2" />
+  </ItemGroup>
+
+  <!-- Link shared source files from the main project -->
+  <ItemGroup>
+    <Compile Include="..\sizoscope\MstatData.cs" Link="MstatData.cs" />
+    <Compile Include="..\sizoscope\MstatData.Cache.cs" Link="MstatData.Cache.cs" />
+    <Compile Include="..\sizoscope\MstatData.Diff.cs" Link="MstatData.Diff.cs" />
+    <Compile Include="..\sizoscope\MstatData.Enumeration.cs" Link="MstatData.Enumeration.cs" />
+    <Compile Include="..\sizoscope\MstatData.Graph.cs" Link="MstatData.Graph.cs" />
+    <Compile Include="..\sizoscope\MstatData.ObjectModel.cs" Link="MstatData.ObjectModel.cs" />
+    <Compile Include="..\sizoscope\InputFileHelper.cs" Link="InputFileHelper.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <None Include="../README.md">
       <Pack>true</Pack>
       <PackagePath>/</PackagePath>
-    </None>
-    <None Include="../sizoscope-cli/bin/Release/net9.0-windows7.0/win-x64/publish/sizoscope-cli.exe">
-      <Pack>true</Pack>
-      <PackagePath>/tools/$(TargetFramework)/any/</PackagePath>
     </None>
   </ItemGroup>
