--- sizoscope-cli/sizoscope-cli.csproj
@@ -1,35 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net9.0-windows7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <RollForward>Major</RollForward>
-    <Version Condition="'$(Version)' == ''">42.42.42.42</Version>
-
-    <PublishAot>true</PublishAot>
-
-    <XmlResolverIsNetworkingEnabledByDefault>false</XmlResolverIsNetworkingEnabledByDefault>
-    <SatelliteResourceLanguages>en-US</SatelliteResourceLanguages>
-    <OptimizationPreference>Size</OptimizationPreference>
-    <StackTraceSupport>false</StackTraceSupport>
-    <UseSystemResourceKeys>true</UseSystemResourceKeys>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="TurboXml" Version="2.0.2" />
-  </ItemGroup>
-
-  <!-- Link shared source files from the main project -->
-  <ItemGroup>
-    <Compile Include="..\sizoscope\MstatData.cs" Link="MstatData.cs" />
-    <Compile Include="..\sizoscope\MstatData.Cache.cs" Link="MstatData.Cache.cs" />
-    <Compile Include="..\sizoscope\MstatData.Diff.cs" Link="MstatData.Diff.cs" />
-    <Compile Include="..\sizoscope\MstatData.Enumeration.cs" Link="MstatData.Enumeration.cs" />
-    <Compile Include="..\sizoscope\MstatData.Graph.cs" Link="MstatData.Graph.cs" />
-    <Compile Include="..\sizoscope\MstatData.ObjectModel.cs" Link="MstatData.ObjectModel.cs" />
-    <Compile Include="..\sizoscope\InputFileHelper.cs" Link="InputFileHelper.cs" />
-  </ItemGroup>
-
-</Project>
